final changes

+++ alpha-mini-rug/alpha_mini_rug/keywords.py
@@ -36,13 +36,13 @@ def key_words(
         str: The keyword found in the user response.
 
     """
-
+    
     global certainty_check
     global answer_found
     global debug_flag
     global key_words_list
-
-    # check if the arguments are of the correct type
+    
+    # # check if the arguments are of the correct type
     if not isinstance(question, str):
         raise TypeError("question is not a string")
     if not isinstance(key_words, list):
@@ -54,7 +54,7 @@ def key_words(
                 raise TypeError("key_words is not a list of strings")
     if not isinstance(time, int):
         raise TypeError("time is not an integer")
-    if not isinstance(certainty, int):
+    if not isinstance(certainty, float):
         raise TypeError("certainty is not an integer")
     if not isinstance(debug, bool):
         raise TypeError("debug is not a boolean")
@@ -62,7 +62,8 @@ def key_words(
     certainty_check = certainty
     debug_flag = debug
     key_words_list = key_words
-
+    
+    
     yield session.call("rie.dialogue.say", text=question, lang=question_lang)
 
     yield session.call("rie.dialogue.config.language", lang=key_words_lang)
@@ -77,7 +78,7 @@ def key_words(
         print("No answer found")
     else:
         print("The keyword found: " + answer_found)
-
+        
     session.call("rie.dialogue.stt.close")
 
     return answer_found
@@ -129,9 +130,9 @@ def key_words_listen(frame):
     ):
         # check the words in the text found
         check_words(frame["data"]["body"]["text"])
-
-        if debug_flag:
-            print("The user input is:")
-            print(frame["data"]["body"]["text"])
+        
+    if debug_flag:
+        print("The user input is:")
+        print(frame["data"]["body"]["text"])
 
     pass
